refactor: separate out presentation actions dialog
