+++ frontend/src/pages/Home.vue
@@ -11,7 +11,7 @@
 				<img src="../icons/slides.svg" class="h-7" />
 				<div class="select-none font-semibold">Slides</div>
 			</div>
-			<Button variant="solid" label="New" size="sm" @click="setDialogProperties('Create')">
+			<Button variant="solid" label="New" size="sm" @click="dialogAction = 'Create'">
 				<template #prefix>
 					<FeatherIcon name="plus" class="h-3.5" />
 				</template>
@@ -112,128 +112,58 @@
 						<Presentation size="16" :strokeWidth="1.5" class="text-white" />
 					</div>
 				</Tooltip>
-				<Tooltip text="Duplicate" :hover-delay="0.3" placement="right">
+				<Tooltip
+					v-for="action in ['Duplicate', 'Rename', 'Delete']"
+					:key="action"
+					:text="action"
+					:hover-delay="0.3"
+					placement="right"
+				>
 					<div
 						class="rounded p-2 mx-2 cursor-pointer bg-gray-200"
-						@click="setDialogProperties('Duplicate')"
+						@click="dialogAction = action"
 					>
-						<Copy size="16" :strokeWidth="1.5" />
-					</div>
-				</Tooltip>
-				<Tooltip text="Rename" :hover-delay="0.3" placement="right">
-					<div
-						class="rounded p-2 mx-2 cursor-pointer bg-gray-200"
-						@click="setDialogProperties('Rename')"
-					>
-						<PenLine size="16" :strokeWidth="1.5" />
-					</div>
-				</Tooltip>
-				<Tooltip text="Delete" :hover-delay="0.3" placement="right">
-					<div
-						class="rounded p-2 mx-2 cursor-pointer bg-gray-200"
-						@click="setDialogProperties('Delete')"
-					>
-						<Trash size="16" :strokeWidth="1.5" />
+						<component
+							:is="
+								action == 'Duplicate' ? Copy : action == 'Rename' ? PenLine : Trash
+							"
+							size="16"
+							:strokeWidth="1.5"
+						/>
 					</div>
 				</Tooltip>
 			</div>
 		</div>
 	</div>
-
-	<Dialog class="pb-0" v-model="showDialog" :options="{ size: 'sm' }">
-		<template #body>
-			<div class="flex flex-col gap-6 p-6">
-				<div class="flex items-center justify-between">
-					<div class="text-md font-semibold text-gray-900">
-						{{ dialogAction }} Presentation
-					</div>
-					<FeatherIcon name="x" class="h-4 cursor-pointer" @click="showDialog = false" />
-				</div>
-				<FormControl
-					v-if="['Rename', 'Duplicate', 'Create'].includes(dialogAction)"
-					:type="'text'"
-					size="md"
-					variant="subtle"
-					placeholder="Presentation Title"
-					v-model="newPresentationTitle"
-				/>
-				<div v-else class="text-base">
-					Are you sure you want to delete this presentation?
-				</div>
-
-				<Button
-					v-if="dialogAction == 'Rename'"
-					variant="solid"
-					label="Update"
-					@click="renamePresentation"
-				>
-					<template #prefix>
-						<FeatherIcon name="edit" class="h-3.5" />
-					</template>
-				</Button>
-
-				<Button
-					v-else-if="dialogAction == 'Duplicate'"
-					variant="solid"
-					label="Create Copy"
-					@click="createPresentation('Duplicate')"
-				>
-					<template #prefix>
-						<FeatherIcon name="copy" class="h-3.5" />
-					</template>
-				</Button>
-
-				<Button
-					v-else-if="dialogAction == 'Delete'"
-					variant="solid"
-					theme="red"
-					label="Confirm Deletion"
-					@click="deletePresentation"
-				>
-					<template #prefix>
-						<FeatherIcon name="trash" class="h-3.5" />
-					</template>
-				</Button>
-
-				<Button v-else variant="solid" label="Create" @click="createPresentation('Create')">
-					<template #prefix>
-						<FeatherIcon name="save" class="h-3.5" />
-					</template>
-				</Button>
-			</div>
-		</template>
-	</Dialog>
+	<PresentationActionDialog :dialogAction="dialogAction" />
 </template>
 
 <script setup>
 import { computed, ref, watch, onBeforeUnmount } from 'vue'
 import { useRouter } from 'vue-router'
-import { createResource, Dialog, FormControl, call, Tooltip } from 'frappe-ui'
-import { Presentation, Copy, PenLine, Trash } from 'lucide-vue-next'
-import { guessTextColorFromBackground } from '../utils/color'
-import tinycolor from 'tinycolor2'
 
+import { createResource, Tooltip } from 'frappe-ui'
+import PresentationActionDialog from '@/components/PresentationActionDialog.vue'
+
+import { Presentation, Copy, PenLine, Trash } from 'lucide-vue-next'
+
+import tinycolor from 'tinycolor2'
 import dayjs from 'dayjs'
 import relativeTime from 'dayjs/plugin/relativeTime'
+
+import { presentationList, activePresentation } from '@/stores/presentation'
 import { startSlideShow } from '@/stores/slide'
+import { guessTextColorFromBackground } from '@/utils/color'
 
 dayjs.extend(relativeTime)
 const router = useRouter()
 
 let interval = null
-const activePresentation = ref(null)
 const previewSlide = ref(0)
 const showLinkToPresentation = ref(false)
 
 const showDialog = ref(false)
 const dialogAction = ref('')
-const newPresentationTitle = ref('')
-
-const presentationList = createResource({
-	url: 'slides.slides.doctype.presentation.presentation.get_all_presentations',
-	method: 'GET',
-	auto: true,
-})
 
 const previewStyles = computed(() => ({
 	backgroundImage: `url(${activePresentation.value?.slides[previewSlide.value]?.thumbnail})`,
@@ -284,65 +214,6 @@ const hidePreview = () => {
 const enablePresentMode = async () => {
 	await router.push(`/${activePresentation.value.name}`)
 	await startSlideShow()
-}
-
-const setDialogProperties = (action) => {
-	dialogAction.value = action
-	if (action == 'Rename') {
-		newPresentationTitle.value = activePresentation.value.title
-	} else if (action == 'Duplicate') {
-		newPresentationTitle.value = `Copy of ${activePresentation.value.title}`
-	} else {
-		newPresentationTitle.value = ''
-	}
-	showDialog.value = true
-}
-
-const resetDialogProperties = () => {
-	showDialog.value = false
-	dialogAction.value = ''
-	newPresentationTitle.value = ''
-}
-
-const createPresentation = async (action) => {
-	resetDialogProperties()
-	let presentation = null
-	if (action == 'Duplicate') {
-		presentation = await call(
-			'slides.slides.doctype.presentation.presentation.duplicate_presentation',
-			{
-				title: newPresentationTitle.value,
-				presentation_name: activePresentation.value.name,
-			},
-		)
-	} else {
-		presentation = await call(
-			'slides.slides.doctype.presentation.presentation.create_presentation',
-			{
-				title: newPresentationTitle.value,
-			},
-		)
-	}
-	await router.push(`/${presentation.name}`)
-}
-
-const renamePresentation = async () => {
-	resetDialogProperties()
-	await call('slides.slides.doctype.presentation.presentation.rename_presentation', {
-		name: activePresentation.value.name,
-		new_name: newPresentationTitle.value,
-	})
-	await presentationList.reload()
-	activePresentation.value.title = newPresentationTitle.value
-}
-
-const deletePresentation = async () => {
-	showDialog.value = false
-	await call('slides.slides.doctype.presentation.presentation.delete_presentation', {
-		name: activePresentation.value.name,
-	})
-	await presentationList.reload()
-	activePresentation.value = null
 }
 
 watch(
